+++ spec/dummy/app/models/komputer.rb
@@ -1,7 +1,7 @@
 class Komputer < ActiveRecord::Base
   self.table_name = "computers"
   store :data, coder: JSON
-  include Storext.model
+  include Storext.model(options: { some: :options })
 
   store_attributes :data do
     manufacturer String, default: "IBM"