+++ spec/dummy/app/models/super_komputer.rb
@@ -1,0 +1,3 @@
+class SuperKomputer < Komputer
+  include Storext::Override
+end