+++ spec/dummy/db/migrate/20160210071144_add_type_to_computers.rb
@@ -1,0 +1,5 @@
+class AddTypeToComputers < ActiveRecord::Migration
+  def change
+    add_column :computers, :type, :text
+  end
+end